requirements.txt
Django==1.6
South==0.8.4
dj-database-url==0.2.2
dj-static==0.0.5
django-debug-toolbar==0.11.0
django-extensions==1.2.5
django-tastypie==0.11.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
python-dateutil==2.2
python-mimeparse==0.1.4
simplejson==3.3.1
six==1.4.1
sqlparse==0.1.10
static==0.4
wsgiref==0.1.2
